Crate linux_syscall_api

This crate provides a safe interface to the Linux syscall API for Starry modules.

Modules

trap

Define the trap handler for the whole kernel

Structs

File

An object providing access to an open file on the filesystem.

FilePath

A struct to represent a file path, which will be canonicalized

OpenFlags

指定文件打开时的权限

Functions

create_link

创建一个链接

init_current_dir

在完成一次系统调用之后，恢复全局目录

new_file

若使用多次new file打开同名文件，那么不同new file之间读写指针不共享，但是修改的内容是共享的

read_file

To read a file with the given path

recycle_user_process

释放所有非内核进程

run_testcase

To run a testcase with the given name and environment variables, which will be used in initproc

Type Aliases

FileFlags

Flags for opening a file

SyscallError

The error of a syscall, which is a LinuxError

SyscallResult

The result of a syscall