Struct linux_syscall_api::OpenFlags

pub struct OpenFlags(/* private fields */);

指定文件打开时的权限

Implementations

impl OpenFlags

pub const RDONLY: OpenFlags = _

只读

pub const WRONLY: OpenFlags = _

只能写入

pub const RDWR: OpenFlags = _

读写

pub const CREATE: OpenFlags = _

如文件不存在，可创建它

pub const EXCLUSIVE: OpenFlags = _

确认一定是创建文件。如文件已存在，返回 EEXIST。

pub const NOCTTY: OpenFlags = _

使打开的文件不会成为该进程的控制终端。目前没有终端设置，不处理

pub const EXCL: OpenFlags = _

同上，在不同的库中可能会用到这个或者上一个

pub const NON_BLOCK: OpenFlags = _

非阻塞读写?(虽然不知道为什么但 date.lua 也要) 在 socket 中使用得较多

pub const TEXT: OpenFlags = _

要求把 CR-LF 都换成 LF

pub const BINARY: OpenFlags = _

和上面不同，要求输入输出都不进行这个翻译

pub const DSYNC: OpenFlags = _

对这个文件的输出需符合 IO 同步一致性。可以理解为随时 fsync

pub const NOFOLLOW: OpenFlags = _

如果是符号链接，不跟随符号链接去寻找文件，而是针对连接本身

pub const CLOEXEC: OpenFlags = _

在 exec 时需关闭

pub const DIR: OpenFlags = _

是否是目录

impl OpenFlags

pub const fn empty() -> OpenFlags

Get a flags value with all bits unset.

pub const fn all() -> OpenFlags

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<OpenFlags>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> OpenFlags

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> OpenFlags

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<OpenFlags>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: OpenFlags) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: OpenFlags) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: OpenFlags)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: OpenFlags)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: OpenFlags)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: OpenFlags, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: OpenFlags) -> OpenFlags

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: OpenFlags) -> OpenFlags

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: OpenFlags) -> OpenFlags

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: OpenFlags) -> OpenFlags

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> OpenFlags

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl OpenFlags

pub const fn iter(&self) -> Iter<OpenFlags>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<OpenFlags>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl OpenFlags

pub fn read_write(&self) -> (bool, bool)

获得文件的读/写权限

pub fn readable(&self) -> bool

获取读权限

pub fn writable(&self) -> bool

获取写权限

pub fn creatable(&self) -> bool

获取创建权限

pub fn new_creatable(&self) -> bool

获取创建新文件权限 与上面的区别是，如果文件已存在，返回 EEXIST

pub fn is_dir(&self) -> bool

获取是否是目录

pub fn is_close_on_exec(&self) -> bool

获取是否需要在 exec() 时关闭